Branch 0.9.0

+++ web/task/src/main/java/com/webank/wedatasphere/qualitis/service/TaskService.java
@@ -32,6 +32,6 @@ public interface TaskService {
      * @return
      * @throws UnExpectedRequestException
      */
-    GeneralResponse<?> getTaskDetail(Integer taskId) throws UnExpectedRequestException;
+    GeneralResponse<?> getTaskDetail(Long taskId) throws UnExpectedRequestException;
 
 }
